Change color order for Rpi5 PIO PWM driver

+++ sources/led-drivers/net/DriverNetLifx.cpp
@@ -1,3 +1,30 @@
+/* DriverNetLifx.cpp
+*
+*  MIT License
+*
+*  Copyright (c) 2020-2025 awawa-dev
+*
+*  Project homesite: https://github.com/awawa-dev/HyperHDR
+*
+*  Permission is hereby granted, free of charge, to any person obtaining a copy
+*  of this software and associated documentation files (the "Software"), to deal
+*  in the Software without restriction, including without limitation the rights
+*  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+*  copies of the Software, and to permit persons to whom the Software is
+*  furnished to do so, subject to the following conditions:
+*
+*  The above copyright notice and this permission notice shall be included in all
+*  copies or substantial portions of the Software.
+
+*  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+*  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+*  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+*  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+*  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+*  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+*  SOFTWARE.
+*/
+
 #include <led-drivers/net/DriverNetLifx.h>
 #include <infinite-color-engine/ColorSpace.h>
 
@@ -236,7 +263,7 @@ QJsonObject DriverNetLifx::discover(const QJsonObject& /*params*/)
 				datagram.resize(socket->pendingDatagramSize());
 				socket->readDatagram(datagram.data(), datagram.size(), &sender, &senderPort);
 
-				if (datagram.size() < sizeof(LifxHeader))
+				if (datagram.size() < static_cast<int>(sizeof(LifxHeader)))
 				{
 					Debug(_log, "The response datagram is too small. Got: %i bytes", static_cast<int>(datagram.size()));
 					continue;
